backend/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1

# Celery & Redis
celery==5.3.6
redis==5.2.1

# Microsoft Graph API
msal==1.31.1
O365==2.0.38
requests==2.32.3

# PDF Processing
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0

# Excel Processing
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

# Fuzzy Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1

# API Documentation
drf-spectacular==0.27.2

# Utilities
python-dateutil==2.9.0
python-decouple==3.8
pytz==2024.2
Unidecode==1.3.8
python-json-logger==2.0.7
django-redis==5.4.0

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.31.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
